Refactor highlight overlay into separate class

+++ GalCon-Client/GalCon-Client/src/com/xxx/galcon/model/Bounds.java
@@ -11,8 +11,8 @@ public class Bounds {
 		this.size = size;
 	}
 
-	public int getTopY() {
-		return (int) origin.y + size.height;
+	public float getTopY() {
+		return origin.y + size.height;
 	}
 
 	public void applyBounds(Actor actor) {
