+++ GalCon-Client/GalCon-Client/src/com/xxx/galcon/model/Size.java
@@ -1,10 +1,10 @@
 package com.xxx.galcon.model;
 
 public class Size {
-	public int width;
-	public int height;
+	public float width;
+	public float height;
 
-	public Size(int width, int height) {
+	public Size(float width, float height) {
 		this.width = width;
 		this.height = height;
 	}
